Add files via upload
Reducing ADC energy using DMA:
The objective was to develop the code for the Leopard Gecko to indicate whether the on board temperature indicates a temperature below a low limit set point or indicates a temperature above another set point. The temperature used to determine whether the temperature has gone outside its temperature settings, will be an average of 750 temperature readings. To save energy by not requiring the CPU to be on during these 750 readings, the results from the ADC should be sent to memory via DMA and only after 750 readings should the CPU wake up, calculate the average, and turn on LED1 if outside of the temperature limit settings.