Add files via upload
I2C Ambient Light Sensor and Load Power Management:
The objective was to develop I2C code for the Leopard Gecko as a master on the I2C bus. The I2C device is the TSL2651 ambient light sensor. For this assignment, the passive on board ambient light sensor will not be used, but the active TSL2651 will be used instead. The LETIMER0 will be used to determine when the TSL2651 should be load power management turned on and when it should be load power management disabled.