Add files via upload
Checked in the code for MCIoT Assignment 6. 
The objective was to develop a system where the Atmel ATSAMB11 is a master controller of a sensor hub, the Leopard Gecko. The ATSAMB11 will take the status of Lightness and Darkness to turn on its LED as well as take the temperature data from the Leopard Gecko and transmit the temperature through its Health Thermometer Service to the Atmel Smart application on a smart phone or tablet.